Renamed method parameters and response object types for actions on Space and User entity.
Fixed KTLINT issues.
