+++ pubnub-entities/src/main/kotlin/com/pubnub/entities/models/consumer/user/PNUserResult.kt
